requirements.txt
Flask
google-genai
python-docx
